requirements.txt
Django==5.2.5
channels==4.2.1
channels-redis==4.2.0
djangorestframework==3.15.2
django-cors-headers==4.4.0
httpx==0.27.2
pydub==0.25.1
python-decouple==3.8
aiofiles==24.1.0
daphne==4.2.1
